==> scrape_data/00_Testing_out_diff_scraping_tools.ipynb <==
# Best way to scrape tables from sref is using pyquery

import requests
from pyquery import PyQuery as pq
import pandas as pd

url = "http://www.basketball-reference.com/players/w/walljo01.html"

response = requests.get(url)

# looks like this helps fix any issues with any scraping library
# can now just use pandas to scrape all dfs on player page
html = response.text
html = html.replace('<!--', '').replace('-->', '')

d = pq(html)
rows = d("#totals > tbody > tr")
headers = d("#totals > thead > tr > th")

rows

# data from each row
data = [[td.text_content() for td in row.iterchildren()] for row in rows]

headers

[i.text_content() for i in headers]

cols = [i.text_content() for i in headers]

df = pd.DataFrame(data=data, columns=cols)

df.head()

## Try pandas

dfs = pd.read_html(html)

len(dfs)

dfs[0].info()

dfs[1].head()

dfs[2].info()

dfs[3].head()

dfs[5].info()

dfs[5].head()

dfs[17].head() # college stats

dfs[19]

dfs[-1] # contract

# Scraping college link

college_link = d("#inner_nav > ul > li:nth-child(8) > div > ul:nth-child(4) > li > a")[0]

college_link.attrib["href"]

college_link.text_content()

### Below is a better way as it doesn't rely on extracting the nth child of lists since that can change depending on the player page

### Instead we rely on extracting all links in the player's inpage nav menu and extracting the one associated to their college basketball stat page

inner_nav_links = d("#inner_nav > ul > li.condensed.hasmore > div > ul > li > a")

inner_nav_links

[i.text_content() for i in inner_nav_links]

list_text = [i.text_content() for i in inner_nav_links]

# get the index of the sublist that contains the substring about college stats
college_text = "College Basketball at Sports-Reference.com"
idx = [i for i, s in enumerate(list_text) if college_text in s]

idx

college_li = inner_nav_links[idx[0]]

college_li.text_content()

college_li.attrib["href"]

# Try scraping links to European stats

jennings_url = "http://www.basketball-reference.com/players/j/jennibr01.html"

response2= requests.get(jennings_url)
html2 = response2.text
html2 = html2.replace('<!--', '').replace('-->', '')

d2 = pq(html2)

inner_nav_links2 = d2("#inner_nav > ul > li.condensed.hasmore > div > ul > li > a")

euro_text = "Euro Stats at Basketball-Reference.com"

list_text2 = [i.text_content() for i in inner_nav_links2]
idx2 = [i for i, s in enumerate(list_text2) if euro_text in s][0]

idx2

d2("#inner_nav > ul > li.condensed.hasmore > div > ul > li > a")

euro_li = inner_nav_links2[idx2]
euro_li.attrib["href"]

# Scraping college stats

Doesn't seem to work except for the very first per game table.

Not sure why, maybe broken html???

Also it's nbd since the bref pages contain the players college stats, plus I found how to access them via the season finder tool.

duncan_url = "http://www.sports-reference.com/cbb/players/tim-duncan-1.html"
response3 = requests.get(duncan_url)
# html3 = response3.text#.replace('<--', '').replace('-->', '')
html3 = response3.content
duncan_pq = pq(html3)

def get_table(pq_obj, table_id_selector):
    rows = pq_obj("{} > tbody > tr".format(table_id_selector))
    headers = pq_obj("{} > thead > tr > th".format(table_id_selector))
    # data from each row
    data = [[td.text_content() for td in row.iterchildren()] for row in rows]
    cols = [i.text_content() for i in headers]
    df = pd.DataFrame(data=data, columns=cols)
    return df

duncan_totals = get_table(duncan_pq, "#players_totals")

duncan_totals

duncan_pq("#players_per_game")

duncan_pq("table")

from bs4 import BeautifulSoup

soup = BeautifulSoup(html3, "lxml")

soup.select("#players_per_game > tbody > tr")[0]

soup.select("#players_totals")

soup2 = BeautifulSoup(html3, "html5lib")

soup2.select("#players_per_game > tbody > tr")[0]

soup2.select("#players_per_game")

soup3 = BeautifulSoup(html3, "html.parser")

len(soup3.find_all("tr"))

len(soup.find_all("table"))

d("#totals > tbody > tr")
