==> scrape_data/01_Scrape_Bref_Player_Ids.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
from urllib.request import urlopen
import string

url = "http://www.basketball-reference.com/players/{}/"

def scrape_data(url):
    """Function to scrape data from Bref's player pages"""
    
    # Use bs4 to get the links for each player in order to extract their 
    # bref ids
    html = urlopen(url)
    soup = BeautifulSoup(html, "lxml")
    players = soup.select("th a")
    links = [player.attrs["href"] for player in players]
    
    # use pandss to easily scrape the table
    df = pd.read_html(url)[0]
    df["Link"] = links
    return df

# get each letter of the alphabet except for x
# which will be used in the url
abc = string.ascii_lowercase.replace("x", "")

dfs = [scrape_data(url.format(i)) for i in abc]

len(dfs)

df = pd.concat(dfs, ignore_index=True)

df.head()

df.to_csv("raw_data/bref_players_06_06_17_raw.csv", index=False)

df["Bref_ID"] = df.Link.str.extract("/.*/.*/(.*)\.", expand=False)

df.head()

df.rename(columns={"Birth Date": "Birth_Date"}, inplace=True)

df.head()

df["Birth_Date"] = pd.to_datetime(df.Birth_Date)

df.head()

# get indicator for Hall of Fame
df["HOF"] = df.Player.str.contains("\*").astype(int)

# now strip the * from the player names
df["Player"] = df.Player.str.strip("\*")

heights = df.Ht.str.split("-", expand=True)

df.Ht[(heights[0].astype(float) * 12 + heights[1].astype(float)).isnull()]

df.loc[2097]

ht_inches = heights[0].astype(float) * 12 + heights[1].astype(float)

df["Ht_Inches"] = ht_inches

df.head()

df.info()

df.to_csv("processed_data/bref_player_info_and_id_06_06_17.csv", index=False)
