==> scrape_data/02_Scrape_Sref_CBB_Player_Ids.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
from urllib.request import urlopen
import string

url = "http://www.sports-reference.com/cbb/players/{}-index.html"

def scrape_data(url):
    html = urlopen(url)
    soup = BeautifulSoup(html, "lxml")
    players = soup.select("#content div p")
    data = [[player.find("a").attrs["href"], player.get_text()] 
            for player in players]
    df = pd.DataFrame(data, columns=["link", "text"])
    return df

abc = string.ascii_lowercase

dfs = [scrape_data(url.format(i)) for i in abc]

len(dfs)

df = pd.concat(dfs, ignore_index=True)

df.head()

df.tail()

df.info()

df[["Player", "Years", "College"]] = df.text.str.split("( \(.*\) )", expand=True)

df.head()

df.Player.values[:10]

df.Years.values[:10]

df.College.values[:10]

df[["link", "text"]].to_csv("raw_data/sref_cbb_player_info_07_06_17_raw.csv", index=False)

df.loc[:, "Years"] = df.Years.str.replace("(\(|\))", "").str.rstrip().str.lstrip()

df.head()

df["From"] = df.Years.str.split("-", expand=True)[0].astype(int)
df["To"] = df.Years.str.split("-", expand=True)[1].astype(int)

df.Years.unique()

df.loc[df.Years=="Gordo Castillo 2011-2011", :]
