==> scrape_data/03_Scraping_college_ids_through_player_pages.ipynb <==
import requests
from pyquery import PyQuery as pq
import pandas as pd

player_df = pd.read_csv("processed_data/bref_player_info_and_id_06_06_17.csv")

player_df.info()

player_df.head()

def scrape_cbb_id(link):
    try:
        url = "http://www.basketball-reference.com" + link
        response = requests.get(url)
        html = response.text.replace('<!--', '').replace('-->', '')
        d = pq(html)
        inner_nav_links = d("#inner_nav > ul > li.condensed.hasmore > div > ul > li > a")
        list_text = [i.text_content() for i in inner_nav_links]
        # get the index of the sublist that contains the substring about 
        # college stats
        college_text = "College Basketball at Sports-Reference.com"
        idx = [i for i, s in enumerate(list_text) if college_text in s][0]
        college_li = inner_nav_links[idx]
        link = college_li.attrib["href"]
        return link
    except Exception as e:
        return [url, e]

test_run = player_df.loc[:5, "Link"].apply(scrape_cbb_id)

test_run

test_run[0]

test_run[1]

cbb_ids = player_df.Link.apply(scrape_cbb_id)

player_df["cbb_info"] = cbb_ids
player_df.to_csv("raw_data/cbb_id_info.csv", index=False)

cbb_ids.apply(type)

player_df.head()

is_list = player_df.cbb_info.apply(isinstance, args=(list,))

len(player_df.loc[is_list, "cbb_info"])

player_df.loc[is_list, "cbb_info"] = ""

player_df2 = player_df.loc[:, ["Bref_ID", "cbb_info"]]

player_df2.rename(columns={"cbb_info": "Sref_Cbb_ID"}, inplace=True)

player_df2["Sref_Cbb_ID"] = player_df2.Sref_Cbb_ID.str.extract("/.*/.*/(.*)\.", 
                                                               expand=False)

player_df2.to_csv("processed_data/sref_cbb_ids_07_07_17.csv", index=False)
