==> scrape_data/04_Create_New_Table_with_name_bref_and_cbb_ids.ipynb <==
import pandas as pd

bref_df = pd.read_csv("processed_data/bref_player_info_and_id_06_06_17.csv")
cbb_df = pd.read_csv("processed_data/sref_cbb_ids_07_07_17.csv")

bref_df.head()

cbb_df.head()

merged_df = pd.merge(bref_df[["Player", "Bref_ID"]], cbb_df, how="left",
                     on="Bref_ID")

merged_df.head()

merged_df.to_csv("processed_data/player_names_bref_and_cbb_ids.csv", index=False)
