==> scrape_data/05_Scrape_Advanced_Cbb_Stats_One_page.ipynb <==
import requests
from pyquery import PyQuery as pq
import pandas as pd

## Play around for one page

url = "http://www.sports-reference.com/cbb/play-index/psl_finder.cgi?request=1&match=single&year_min=1996&year_max=&conf_id=&school_id=&class_is_fr=Y&class_is_so=Y&class_is_jr=Y&class_is_sr=Y&pos_is_g=Y&pos_is_gf=Y&pos_is_fg=Y&pos_is_f=Y&pos_is_fc=Y&pos_is_cf=Y&pos_is_c=Y&games_type=A&qual=&c1stat=&c1comp=&c1val=&c2stat=&c2comp=&c2val=&c3stat=&c3comp=&c3val=&c4stat=&c4comp=&c4val=&order_by=ws&order_by_asc=&offset=96700"

response = requests.get(url)

# looks like this helps fix any issues with any scraping library
# can now just use pandas to scrape all dfs on player page
html = response.text
html = html.replace('<!--', '').replace('-->', '')

d = pq(html)
rows = d("#stats tbody > tr")

rows

type(rows[0])

header_drop = rows[21]

header_drop.attrib

rows[20].attrib

rows[0].attrib

rows[20].attrib == ""

data = [[td.text_content() for td in row.iterchildren()] for row in rows]

len(data)

data2 = [[td.text_content() for td in row.iterchildren()] 
         for row in rows if row.attrib == ""]

len(data2)

first_row = rows[0]

player = first_row.getchildren()[1]

player.getchildren()[0].attrib["href"]

def get_cbb_link_and_data(idx, td):
    """Helps get the player link as you scrape the data"""
    if idx == 1:
        return td.text_content(), player.getchildren()[0].attrib["href"]
    else:
        return td.text_content()

data3 = [[get_cbb_link_and_data(idx, td) 
          for idx, td in enumerate(row.iterchildren())] 
         for row in rows if row.attrib == ""]

len(data3)

data3[:2]

poptip = pq(".poptip")

poptip.text

headers = d("#stats > thead > tr > th")

headers

[th.text_content() for th in headers[2:]]

def get_table(pq_obj, table_id_selector):
    rows = pq_obj("{} > tbody > tr".format(table_id_selector))
    headers = pq_obj("{} > thead > tr > th".format(table_id_selector))
    # data from each row
    data = [[get_cbb_link_and_data(idx, td) 
             for idx, td in enumerate(row.iterchildren())] 
            for row in rows if row.attrib == ""]
    cols = [th.text_content() for th in headers[2:]]
    df = pd.DataFrame(data=data, columns=cols)
    return df

df = get_table(d, "#stats")

df.head()
