==> scrape_data/07_Scrape_All_Traditional_Cbb_Stats.ipynb <==
import requests
from pyquery import PyQuery as pq
import pandas as pd

url_template = "http://www.sports-reference.com/cbb/play-index/psl_finder.cgi"\
               "?request=1&match=single&year_min=&year_max=&conf_id=&school"\
               "_id=&class_is_fr=Y&class_is_so=Y&class_is_jr=Y&class_is_sr=Y"\
               "&pos_is_g=Y&pos_is_gf=Y&pos_is_fg=Y&pos_is_f=Y&pos_is_fc=Y&"\
               "pos_is_cf=Y&pos_is_c=Y&games_type=A&qual=&c1stat=&c1comp=&c1"\
               "val=&c2stat=&c2comp=&c2val=&c3stat=&c3comp=&c3val=&c4stat=&c4"\
               "comp=&c4val=&order_by=pts&order_by_asc=&offset={}"

def get_cbb_link_and_data(idx, td):
    """Helps get the player link as you scrape the data""" 
    
    if idx == 1:
        return td.text_content(), td.getchildren()[0].attrib["href"]
    else:
        return td.text_content()

def create_pq(url):
    """Creates PyQuery object used for scraping"""
    
    response = requests.get(url)
    html = response.text.replace('<!--', '').replace('-->', '')
    return pq(html)

def get_table(url, table_id_selector):
    """Scrapes the data and returns it as a DataFrame"""
    
    pq_obj = create_pq(url)
    
    rows = pq_obj("{} > tbody > tr".format(table_id_selector))
    headers = pq_obj("{} > thead > tr > th".format(table_id_selector))
    
    # get data from each row and coll
    data = [[get_cbb_link_and_data(idx, td) 
             for idx, td in enumerate(row.iterchildren())] 
            for row in rows if row.attrib == ""]
    cols = [th.text_content() for th in headers[2:]]
    df = pd.DataFrame(data=data, columns=cols)
    
    return df

def scrape_data(offset):
    url = url_template.format(offset)
    try:
        return get_table(url, "#stats")
    except Exception as e:
        return [url, e]

dfs = [scrape_data(offset) for offset in range(0, 108601, 100)]

len(dfs)

errors = [error for error in dfs if isinstance(error, list)]

len(errors)

dfs[0].head()

actual_dfs = [df for df in dfs if isinstance(df, pd.DataFrame)]

len(actual_dfs)

df = pd.concat(actual_dfs, ignore_index=True)

df.head()

df.tail()

df.info()

df.to_csv("raw_data/cbb_player_season_trad_stats_07_08_17_raw.csv",
          index=False)

errors

errors[0][0]
