==> scrape_data/08_Scrape_Page_that_gives_error.ipynb <==
import requests
from pyquery import PyQuery as pq
import pandas as pd

url = 'http://www.sports-reference.com/cbb/play-index/psl_finder.cgi?request=1&match=single&year_min=&year_max=&conf_id=&school_id=&class_is_fr=Y&class_is_so=Y&class_is_jr=Y&class_is_sr=Y&pos_is_g=Y&pos_is_gf=Y&pos_is_fg=Y&pos_is_f=Y&pos_is_fc=Y&pos_is_cf=Y&pos_is_c=Y&games_type=A&qual=&c1stat=&c1comp=&c1val=&c2stat=&c2comp=&c2val=&c3stat=&c3comp=&c3val=&c4stat=&c4comp=&c4val=&order_by=pts&order_by_asc=&offset=105800'

# Error comes about when trying to scrape player cbb links
### Will scrape data an links/ids separately.

response = requests.get(url)
html = response.text.replace('<!--', '').replace('-->', '')
d = pq(html)

rows = d("#stats > tbody > tr")
headers = d("#stats > thead > tr > th")

data = [[td.text_content() for td in row.iterchildren()]
        for row in rows if row.attrib == ""]
cols = [th.text_content() for th in headers[2:]]
df = pd.DataFrame(data=data, columns=cols)

df.head()

## Realized issues aris from accessing attributes using the "href" key, should have used the get() method in original function

[row[1].attrib for row in rows]

# also realized this is a better method of getting the cbb id, instead
# of having to scrape the link and extract it from there, oh well
ids = [row[1].attrib.get("data-append-csv") for row in rows if row.attrib == ""]

ids[:10]

df["Sref_Cbb_ID"] = ids

df.head()

df.tail()

df.to_csv("raw_data/error_df_trad_stats_07_09_17.csv", index=False)
