==> scrape_data/09_Clean_College_Stats.ipynb <==
import pandas as pd

trad_df = pd.read_csv("raw_data/cbb_player_season_trad_stats_07_08_17_raw.csv")
adv_df = pd.read_csv("raw_data/cbb_player_season_adv_stats_07_08_17_raw.csv")

trad_df.head()

trad_df.info()

adv_df.head()

adv_df.info()

adv_players = adv_df.Player.str.split(', ', expand=True)

adv_players[0] = adv_players[0].str.strip("(\'")

adv_players[1] = adv_players[1].str.strip(")\'").str.extract("/.*/.*/(.*)\.", 
                                                             expand=False)

adv_players.head()

adv_players.values[:2]

adv_df[["Player", "Sref_Cbb_ID"]] = adv_players[[0, 1]]

adv_df.head()

trad_players = trad_df.Player.str.split(', ', expand=True)
trad_players[0] = trad_players[0].str.strip("(\'")
trad_players[1] = trad_players[1].str.strip(")\'").str.extract("/.*/.*/(.*)\.", 
                                                             expand=False)
trad_players.head()

trad_df[["Player", "Sref_Cbb_ID"]] = trad_players[[0, 1]]

trad_df.head()

error_trad_df = pd.read_csv("raw_data/error_df_trad_stats_07_09_17.csv")

error_trad_df.head()

error_trad_df.info()

trad_df.info()

trad_df2 = pd.concat([trad_df, error_trad_df], ignore_index=True)

trad_df2.info()

adv_df.info()

## Clean up column names

adv_df.columns = adv_df.columns.str.replace("%", "_Pct")

trad_df2.columns.values[11:15] = ["FG_" + col for col in 
                                  trad_df2.columns.values[11:15]]

adv_df.columns

trad_df2.columns

## Create some FG% columns

trad_df2["FG_Pct"] = trad_df2.FG / trad_df2.FGA
trad_df2["FG_2P_Pct"] = trad_df2.FG_2P / trad_df2.FG_2PA
trad_df2["FG_3P_Pct"] = trad_df2.FG_3P / trad_df2.FG_3PA

trad_df2.head()

# almost forgot FT%
trad_df2["FT_Pct"] = trad_df2.FT / trad_df2.FTA

trad_df2.head()

trad_df2.head()

## Drop Rk column

trad_df2.drop("Rk", axis=1, inplace=True)
adv_df.drop("Rk", axis=1, inplace=True)

trad_df2.info()

adv_df.info()

adv_df.to_csv("processed_data/cbb_player_seasons_adv_stats_07_09_17.csv",
              index=False)
trad_df2.to_csv("processed_data/cbb_player_season_trad_stats_07_09_17.csv",
                index=False)
