==> scrape_data/10_Scrape_Draft_Express_Basic_Stats_Single_Page.ipynb <==
import requests
from pyquery import PyQuery as pq
import pandas as pd

url = "http://www.draftexpress.com/stats/ncaa/2017/all/basic/pace/0/all/all/1/ptspg/desc"

# type out cols just cuz the dx col headers are a mess
cols = ["box", "Player", "Team_Logo", "Team", "G", "MP", "PTS", "FG_2P", 
        "FG_2PA",  "FG_2P_Pct", "FG_3P", "FGA_3P", "FG_3P_Pct", "FT", "FTA",
        "FT_Pct", "OREB", "DREB", "REB", "AST", "STL", "BLK", "TO", "PF"]

response = requests.get(url)
html = response.text
d = pq(html)

row_selector = "#cmn_wrap > div.row.two-cols > div > div.row.inner-page.stats > div > table > tbody > tr"

rows = d(row_selector)

rows

first_row = rows[0]

[td.text_content() for td in first_row.iterchildren()]

data = [[td.text_content() for td in row] for row in rows]

len(data)

data[:2]

for idx, i in enumerate(data[0]): print(idx, i)

df = pd.DataFrame(data, columns=cols)

for row in data: print(len(row))

url_selector= "#cmn_wrap > div.row.two-cols > div >"\
              "div.row.inner-page.stats > div > table > tbody > tr > "\
              "td.text.key > a"
urls = d(url_selector)

urls[0].attrib["href"]

len(rows)

len(urls)

urls[0].get("href")

def get_links(pq_obj, url_selector):
    urls = pq_obj(url_selector)
    links = [url.get("href") for url in urls]
    return links

team_url_selector = "#cmn_wrap > div.row.two-cols > div > "\
                    "div.row.inner-page.stats > div > table > tbody > tr > td"\
                    " > a"
team_links = get_links(d, team_url_selector)

#cmn_wrap > div.row.two-cols > div > div.row.inner-page.stats > div > table > tbody > tr:nth-child(1) > td:nth-child(4) > a

team_links[:10]

player_links = get_links(d, url_selector)

player_links[:10]

team_url_selector2 = ".key~ .text+ td a"
team_links = get_links(d, team_url_selector2)

team_links[:10]

team_url_selector3 = "#cmn_wrap > div.row.two-cols > div > div.row.inner-page.stats > div > table > tbody > tr > td.highlighted > a"
team_links2 = get_links(d, team_url_selector3)

team_links2[:10]

last_pg_selector = ".disabled+ li a"
last_pg = d(last_pg_selector)

last_pg

last_pg[0].text_content()
