==> scrape_data/13_Scrape_Draft_Express_Efficiency_Player_Stats.ipynb <==
import requests
import pandas as pd
from pyquery import PyQuery as pq

url_template = "http://www.draftexpress.com/stats/ncaa/{year}/all/efficiency/"\
                "pace/0/all/all/{pg}"

def create_pq(url):
    """Creates PyQuery object used for scraping"""
    response = requests.get(url)
    html = response.text
    return pq(html)

def get_last_pg(url):
    """Get the last page number to be scraped"""
    pq_obj = create_pq(url)
    last_pg_selector = ".disabled+ li a"
    last_pg = pq_obj(last_pg_selector)
    return last_pg[0].text_content()

def get_links(pq_obj, url_selector):
    """Gets the links associated with the given css selector"""
    urls = pq_obj(url_selector)
    links = [url.get("href") for url in urls]
    return links

def get_data(pq_obj, row_selector):
    """Get table data"""
    rows = pq_obj(row_selector)
    data = [[td.text_content() for td in row] for row in rows]
    return data

def create_df(url, cols):
    """Scrapes data from url and returns a DataFrame with given columns"""
    pq_obj = create_pq(url)
    
    # Extract the links for the players and teams
    player_selector = "#cmn_wrap > div.row.two-cols > div >"\
                      "div.row.inner-page.stats > div > table > tbody > tr > "\
                      "td.text.key > a"
    player_links = get_links(pq_obj, player_selector)
    team_selector = ".key~ .text+ td a"
    team_links = get_links(pq_obj, team_selector)
    
    # get the table data
    row_selector = "#cmn_wrap > div.row.two-cols > div > "\
                   "div.row.inner-page.stats > div > table > tbody > tr"
    data = get_data(pq_obj, row_selector)

    df = pd.DataFrame(data=data, columns=cols)
    df["Player_Link"] = player_links
    df["Team_Link"] = team_links
    
    return df

# type out cols just cuz the dx col headers are a mess
cols = ["box", "Player", "Team_Logo", "Team", "G", "MP", "PTS", "FGA", 
        "PTS_per_Play", "TS_Pct", "eFG_Pct", "FT_Rate", "Three_Pt_Rate",
        "AST", "AST_to_FGA_Ratio", "AST_to_TO_Ratio", "PPR", "STL", "BLK",
        "PF"]

len(cols)

last_pgs = {yr: int(get_last_pg(url_template.format(year=yr, pg=1)))
            for yr in range(2003,2018)}

last_pgs

dfs = []
errors = []

for yr in range(2003, 2018):
    for pg in range(1, last_pgs.get(yr)):
        url = url_template.format(year=yr, pg=pg)
        try:
            df = create_df(url, cols)
            df["Season"] = yr
            dfs.append(df)
        except Exception as e:
            errors.append([url, e])

len(dfs)

len(errors)

errors

df = pd.concat(dfs, ignore_index=True)

df.info()

df.head()

df.tail()

df["DX_Player_ID"] = df.Player_Link.str.extract("/.*/(.*)/", expand=False)
df["DX_College_ID"] = df.Team_Link.str.extract("/.*/.*/.*/(.*)", expand=False)
df.head()

df.to_csv("raw_data/draft_express_player_efficiency_stats_07_10_17.csv", 
          index=False)
