==> scrape_data/15_Scrape_Draft_Express_RSCI_Single_Page_SCRAP.ipynb <==
import requests
import pandas as pd
from pyquery import PyQuery as pq

url = "http://www.draftexpress.com/RSCI/1998/"
html = requests.get(url).text
d = pq(html)

headers = d("th")

headers

[th.text_content() for th in headers]

rows = d("tr")

rows

len(rows)

rows[0].text_content()

from numpy import s_

[td.text_content() for td in first_row[:2]] +
[td.text_content() for td in first_row[-3:-1]]

player_selector = ".key:nth-child(2) a"
d(player_selector)

len(rows)

len(d(player_selector))

first_row = rows[1]

first_row.getchildren()[s_[0:2, 4]]

[td.text for td in rows[1][:3]]

len(rows[1])

[td.text for td in rows[4]]
