==> scrape_data/18_Scrape_Draft_Express_European_Player_Stats.ipynb <==
import requests
import pandas as pd
from pyquery import PyQuery as pq
import helpers

leagues = ["euroleague", "eurocup", "acb", "italian", "french", "greek"]
stats = ["basic", "usage", "efficiency"]

url_template = "http://www.draftexpress.com/stats/{league}/{year}/all/{stat}/"\
                "pace/0/all/all/{pg}"

# dict to store scraped dfs and errors for each league and stat type
data_dict = {lg: {stat: {"dfs": [], "errors": []} for stat in stats} 
             for lg in leagues}

data_dict

col_dict = {"basic": helpers.basic_stats_cols,
            "usage": helpers.usage_cols,
            "efficiency": helpers.eff_cols}

print(col_dict)

for lg in leagues:
    for stat in stats:
        for yr in range(2003, 2018):
            # For each year find the last page number, if there isn't
            # one then there are no stats for that page and move on to a 
            # different year
            last_pg_url = url_template.format(league=lg, year=yr, stat=stat, 
                                              pg=1)
            try:
                last_pg = helpers.get_last_pg(last_pg_url)
            except Exception as e:
                data_dict[lg][stat]["errors"].append([lg, stat, yr,
                                                      last_pg_url, e])
                continue
            for pg in range(1, int(last_pg)):
                url = url_template.format(league=lg,year=yr,stat=stat, pg=pg)
                try:
                    df = helpers.create_df(url, col_dict[stat])
                    df["Season"] = yr
                    df["League"] = lg
                    df.drop(["box", "Team_Logo"], axis=1, inplace=True)
                    df["DX_Player_ID"] = (df.Player_Link.str
                                          .extract("/.*/(.*)/", expand=False))
                    df["DX_Team_ID"] = (df.Team_Link.str
                                        .extract("/.*/.*/.*/(.*)", expand=False))
                    df["Team"] = df.Team.str.strip()
                    data_dict[lg][stat]["dfs"].append(df)
                except Exception as e:
                    data_dict[lg][stat]["errors"].append([lg, stat, yr, url, 
                                                          e])

for lg in leagues:
    for stat in stats:
        print(len(data_dict[lg][stat]["errors"]))

for lg in leagues:
    for stat in stats:
        print(len(data_dict[lg][stat]["dfs"]))

for lg in leagues:
    for stat in stats:
        print(data_dict[lg][stat]["errors"],"\n")

for lg in leagues:
    for stat in stats:
        df = pd.concat(data_dict[lg][stat]["dfs"])
        if stat == "basic":
            file = "raw_data/draft_express_{}_player_basic_stats_pace_adj_"\
                   "07_11_17.csv"
            df.to_csv(file.format(lg), index=False)
        else:
            file = "raw_data/draft_express_{}_player_{}_07_11_17.csv"
            df.to_csv(file.format(lg, stat), index=False)
