==> scrape_data/19_Clean_Up_Draft_Express_Data.ipynb <==
import pandas as pd

basic_adj_df = pd.read_csv("raw_data/draft_express_player_basic_stats_pace_adj_07_10_17.csv")
basic_pg_df = pd.read_csv("raw_data/draft_express_player_basic_stats_per_game_07_10_17.csv")
usage_df = pd.read_csv("raw_data/draft_express_player_usage_stats_07_10_17.csv")
eff_df = pd.read_csv("raw_data/draft_express_player_efficiency_stats_07_10_17.csv")
# rsci_df = pd.read_csv("raw_data/draft_express_player_RSCI_ranks_07_10_17.csv")
# meas_df = pd.read_csv("raw_data/draft_express_player_measurements_07_10_17.csv")

pd.options.display.max_columns = None

basic_adj_df.head()

basic_pg_df.head()

eff_df.head()

usage_df.head()

basic_adj_df.drop(['box', 'Team_Logo'], axis=1, inplace=True)
basic_pg_df.drop(['box', 'Team_Logo'], axis=1, inplace=True)
eff_df.drop(['box', 'Team_Logo'], axis=1, inplace=True)
usage_df.drop(['box', 'Team_Logo'], axis=1, inplace=True)

basic_adj_df['Team'] = basic_adj_df.Team.str.strip()
basic_pg_df['Team'] = basic_pg_df.Team.str.strip()
eff_df['Team'] = eff_df.Team.str.strip()
usage_df['Team'] = usage_df.Team.str.strip()

basic_pg_df.head()

basic_pg_df['FG_2P_Pct'] = basic_pg_df.FG_2P_Pct.str.replace('%', '').astype(float) / 100
basic_pg_df['FG_3P_Pct'] = basic_pg_df.FG_3P_Pct.str.replace('%', '').astype(float) / 100


basic_adj_df['FG_2P_Pct'] = basic_adj_df.FG_2P_Pct.str.replace('%', '').astype(float) / 100
basic_adj_df['FG_3P_Pct'] = basic_adj_df.FG_3P_Pct.str.replace('%', '').astype(float) / 100


basic_pg_df['FT_Pct'] = basic_pg_df.FT_Pct.str.replace('%', '').astype(float) / 100
basic_adj_df['FT_Pct'] = basic_adj_df.FT_Pct.str.replace('%', '').astype(float) / 100

basic_adj_df.to_csv('processed_data/draft_express_player_basic_stats_pace_adj_2017-08-24.csv', 
                    index=False)
basic_pg_df.to_csv('processed_data/draft_express_player_basic_stats_per_game_2017-08-24.csv', 
                    index=False)
usage_df.to_csv('processed_data/draft_express_player_usage_stats_2017-08-24.csv', 
                index=False)
eff_df.to_csv('processed_data/draft_express_player_efficiency_stats_2017-08-24.csv', 
              index=False)
